PHOENIX-5827 Add a maven repo to PQS and optionally bundle phoenix-client
* Move the cluster.xml to the prescribed location per Maven convention.
* Use maven-assembly-plugin to create a Maven repo in the assembly
* Pull the phoenix-client jar from central and localize so PQS can actually function
* Fix the phoenix-client jar name so that it's picked up by phoenix_utils.py
* Prevent old Jetty versions from sneaking in on HBase 1.x
* Build the ServerCustomizer and expose configuration to enable it

Move integration tests to a dedicated module to avoid jetty clashing

With Hadoop2/HBase1, we have to deal with conflicting versions of Jetty
at runtime. Avatica's ServerCustomizers expose an unshaded Jetty class
(Server), which causes problems when we have Jetty6 also on the
classpath.

We can avoid this by moving all things that touch Avatica's version of
Jetty into queryserver, shade that, and then only invoke HBase/Hadoop
from within a different module (letting their Jetty6 run wild).

The only gripe is that BasicAuthenticationServerCustomizer has to live
in src/main/java to get shaded, even though it is test-only code. This
is a minor grievance for what's a horrible runtime solution. This all
gets much better with Hadoop3/HBase2 where Jetty is shaded.

Closes #25

Signed-off-by: Istvan Toth <stoty@apache.org>